The `hasAttributes` trait split has been canceled
